+++ src/test/java/amazonTest/AmazonLogin.java
@@ -1,6 +1,7 @@
 package amazonTest;
 
 import amazonHelper.BrowserInit;
+import amazonHelper.DataPro;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
@@ -27,7 +28,7 @@ public class AmazonLogin {
         lpo = new LoginPageObject(driver);
     }
 
-    @Test(groups = {"test1"},dataProvider ="input1" )
+    @Test(groups = {"test1"},dataProvider ="input1" , dataProviderClass = DataPro.class)
     public void validateLogin1(String username, String password) throws InterruptedException {
         driver.get(url);
         Thread.sleep(2000);
@@ -44,7 +45,7 @@ public class AmazonLogin {
         lpo.click_Signin();
     }
 
-    @Test(groups = {"test2"},dataProvider ="input2" )
+    @Test(groups = {"test2"},dataProvider ="input2" , dataProviderClass = DataPro.class )
     public void validateLogin2(String username, String password) throws InterruptedException {
         driver.get(url);
         Thread.sleep(2000);
@@ -61,20 +62,6 @@ public class AmazonLogin {
         lpo.click_Signin();
     }
 
-    @DataProvider(name="input1")
-    public Object[][] getData(){
-        return new Object[][]{
-                {"herainpandey@gmail.com","amazon@1988"},
-                /*{"9717661012","himu@1988"}*/  //   If uncommented this line, then test will run two times
-        };
-    }
-
-    @DataProvider(name="input2")
-    public Object[][] getData1(){
-        return new Object[][]{
-                {"9717661012","himu@1988"}
-        };
-    }
 
     @AfterClass(alwaysRun = true)
     public void close(){
